加了一点代码 懒得写简介了

+++ ghost-dao/src/main/java/cc/w0rm/ghost/mysql/dao/CommodityDALImpl.java
@@ -60,7 +60,7 @@ public class CommodityDALImpl {
         }
         try {
             MsgGroup msgGroup = new MsgGroup();
-            msgGroup.setGroup(Integer.parseInt(groups));
+            msgGroup.setGroup(Long.parseLong(groups));
             msgGroup.setCommodityId(commodityId);
             msgGroup.setInsertTime(String.valueOf(System.currentTimeMillis()));
             msgGroupMapper.insertOrUpdate(msgGroup);
